+++ packages/happn-3/lib/services/cache/cache-persist.js
@@ -1,5 +1,6 @@
 module.exports = class CachePersist extends require('./cache-static') {
   #synced = false;
+  #syncing = false;
   #dataStore;
   #basePath;
   constructor(name, opts) {
@@ -15,6 +16,14 @@ module.exports = class CachePersist extends require('./cache-static') {
       throw new Error(`no dataStore defined for persisted cache on path: ${this.#basePath}`);
     }
     this.#dataStore = opts.dataStore;
+  }
+
+  get isSynced() {
+    return this.#synced;
+  }
+
+  get isSyncing() {
+    return this.#syncing;
   }
 
   get(key, opts = {}, callback) {
@@ -46,7 +55,7 @@ module.exports = class CachePersist extends require('./cache-static') {
       callback = opts;
       opts = {};
     }
-    if (!this.#checkSynced(callback)) {
+    if (!this.#syncing && !this.#checkSynced(callback)) {
       return;
     }
 
@@ -54,7 +63,7 @@ module.exports = class CachePersist extends require('./cache-static') {
     if (!opts.ttl) opts.ttl = this.opts.defaultTTL;
     if (opts.noPersist) {
       const result = super.set(key, data, opts);
-      callback(null, result);
+      return callback(null, result);
     }
 
     this.#persistData(key, this.createCacheItem(key, data, opts), (e) => {
@@ -126,11 +135,16 @@ module.exports = class CachePersist extends require('./cache-static') {
   }
 
   sync(callback) {
+    this.#syncing = true;
     this.#dataStore.get(`${this.#basePath}/*`, (e, items) => {
-      if (e) return callback(e);
+      if (e) {
+        this.#syncing = false;
+        return callback(e);
+      }
 
       if (!items || items.length === 0) {
         this.#synced = true;
+        this.#syncing = false;
         return callback(null);
       }
 
@@ -142,7 +156,6 @@ module.exports = class CachePersist extends require('./cache-static') {
               return this.#removeData(item.data.key, itemCB);
             }
           }
-
           this.set(
             item.data.key,
             item.data.data,
@@ -154,6 +167,7 @@ module.exports = class CachePersist extends require('./cache-static') {
           );
         },
         (e) => {
+          this.#syncing = false;
           if (e) return callback(e);
           this.#synced = true;
           callback();