+++ packages/happner-2/RELEASES.md
@@ -957,6 +957,10 @@ SMC-1645: made error messages more consistent, encapsulating mesh.component.meth
   - updates to commons libs, lint fixes, due to:
   - TEN-49: add diagnostics log to caching service, refactored happn-3 caching layer
   - TEN-54: all systems tested on node v18
+
+12.1.6 2022-06-02
+-----------------
+  - TEN-49: fixed persisted cache and loki provider issues, caused by revoked tokens